# database/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package goAPI

import (
	"time"

	"github.com/google/uuid"
)

type Person struct {
	ID             uuid.UUID `example:"00000000-0000-0000-0000-000000000000" format:"uuid" json:"id"`
	Surname        *string   `example:"Иванов" json:"surname"`
	Name           *string   `example:"Иван" json:"name"`
	Patronymic     *string   `example:"Иванович" json:"patronymic"`
	Address        *string   `example:"3-й Автозаводский проезд, вл13, Москва, 115280" json:"address"`
	PassportSerie  string    `example:"1234" json:"passportSerie" maxLength:"4" minLength:"4"`
	PassportNumber string    `example:"123456" json:"passportNumber" maxLength:"6" minLength:"6"`
}

type Seed struct {
	Name *string `json:"name"`
}

type Task struct {
	ID        uuid.UUID  `example:"00000000-0000-0000-0000-000000000000" format:"uuid" json:"id"`
	PeopleID  uuid.UUID  `json:"peopleId"`
	Name      string     `example:"Помыть посуду" json:"name"`
	StartTime time.Time  `example:"2022-01-01T00:00:00Z" format:"dateTime" json:"startTime"`
	EndTime   *time.Time `example:"2022-01-01T00:00:00Z" format:"dateTime" json:"endTime"`
}